requirements.txt

google-cloud-aiplatform
google-cloud-storage
google-cloud-logging
google-adk[eval]
requests
gcsfs
pypdf
matplotlib 
numpy
pandas
scikit-learn
fsspec
tabulate
PyYAML
